Tracked down and fixed a bug with the ram (it was hooked up improperly).
It was ignoring the cyc and stb signals when hooking up the write enables for the smaller ram granuels.  As a result they were writing when they weren't suppossed to, and overwriting correct data whith garbage.

Also, the ram.ld wasn't using all of the available bootrom.

And I've begun converting constant variables (which use ram) to #defines (which don't, and make syntax nicer).

Also, building will now create an assembly file for every c and cpp file.